enterprise-doc-intelligence/enterprise-doc-intelligence/requirements.txt
# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ── RAG / LangChain ────────────────────────────────────────────────────────────
langchain==0.2.6
langchain-community==0.2.6
langchain-groq==0.1.6

# ── Embeddings ─────────────────────────────────────────────────────────────────
sentence-transformers==3.0.1
torch==2.3.1
transformers==4.42.3

# ── Vector Store ───────────────────────────────────────────────────────────────
faiss-cpu==1.8.0

# ── LLM Provider ───────────────────────────────────────────────────────────────
groq==0.9.0

# ── Document Loaders ───────────────────────────────────────────────────────────
pypdf==4.2.0
python-docx==1.1.2
openpyxl==3.1.5
unstructured==0.14.9
python-pptx==0.6.23

# ── OCR ────────────────────────────────────────────────────────────────────────
pytesseract==0.3.13
Pillow==10.4.0
opencv-python-headless==4.10.0.84

# ── Frontend ───────────────────────────────────────────────────────────────────
streamlit==1.36.0
requests==2.32.3

# ── Config & Utilities ─────────────────────────────────────────────────────────
pydantic==2.8.0
pydantic-settings==2.3.4
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.4.1

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0

# ──────────────────────────────────────────────────────────────────────────────
# INSTALL NOTE (CPU-only PyTorch on Windows):
#   pip install torch==2.3.1 --index-url https://download.pytorch.org/whl/cpu
#   pip install -r requirements.txt
# ──────────────────────────────────────────────────────────────────────────────
